Rename user stores to make difference clearer
One store is used for the logged in user, the other store is
used for all users.
