+++ frontend/src/pages/Profile.svelte
@@ -3,38 +3,46 @@
     import EnvelopeIcon from "svelte-icons/fa/FaEnvelope.svelte";
     import Page from "../layout/Page.svelte";
     import { _ } from "../services/i18n";
-    import { userStore } from "../stores/user";
+    import { loggedInUserStore } from "../stores/loggedInUserStore";
 </script>
 
 <Page editHandler={() => console.log("edit")} title={$_("app.profile.title")}>
-    {#if $userStore}
+    {#if $loggedInUserStore}
         <header>
             <img
-                src={$userStore.image}
+                src={$loggedInUserStore.image}
                 alt={$_("app.profile.imageAltText")}
                 class="userImage atRight"
             />
-            <h3>{$userStore.firstName} {$userStore.lastName}</h3>
+            <h3>
+                {$loggedInUserStore.firstName}
+                {$loggedInUserStore.lastName}
+            </h3>
 
             <p>
                 {$_("app.profile.projects", {
-                    values: { count: $userStore.projectsShortInfo.length },
+                    values: {
+                        count: $loggedInUserStore.projectsShortInfo.length,
+                    },
                 })}
             </p>
 
             <p>
                 {$_("app.profile.registration", {
                     values: {
-                        date: $date(new Date($userStore.registrationDate), {
-                            format: "medium",
-                        }),
+                        date: $date(
+                            new Date($loggedInUserStore.registrationDate),
+                            {
+                                format: "medium",
+                            }
+                        ),
                     },
                 })}
             </p>
         </header>
 
         <p class="fullWidth">
-            {$userStore.description}
+            {$loggedInUserStore.description}
         </p>
 
         <button
@@ -46,7 +54,7 @@
         </button>
 
         <ul class="tag-list">
-            {#each $userStore.tags as tag}
+            {#each $loggedInUserStore.tags as tag}
                 <li>{tag}</li>
             {/each}
         </ul>
