+++ frontend/src/stores/loggedInUserStore.ts
@@ -1,4 +1,4 @@
 import { writable } from 'svelte/store';
 import type { CurrentUserDTO } from '../../../mock-backend/mocks/models/user';
 
-export const userStore = writable<CurrentUserDTO | null>();
+export const loggedInUserStore = writable<CurrentUserDTO | null>();
